Add files via upload
Trying to check numpy.fft.fftn which is a function that should fourier transform an N-dimensional space. Testing by making a gaussian and then fourier transforming it. The result should be another gaussian but instead I'm getting just a big blob/rectangular prism. I fourier transform and shift it but yeah... still a blob.